Rename `input_debug_hud.py` to `debug_hud.py` to match its contents
